# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces": {
      "post": {
        "operationId": "WorkspaceEntry_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "8–64 characters of [A-Za-z0-9_-], unique per intended operation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a workspace owned by the caller (needs an admin password)",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/uploads/workspace-icon": {
      "post": {
        "operationId": "WorkspaceEntry_iconTicket",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTicketDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A presigned upload for a workspace icon (PNG, JPEG or WebP, at most 1 MB)",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/invitations/accept": {
      "post": {
        "operationId": "WorkspaceEntry_accept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInvitationResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a workspace with an invitation link",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}": {
      "get": {
        "operationId": "Workspace_get",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "patch": {
        "operationId": "Workspace_update",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "delete": {
        "operationId": "Workspace_remove",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete the workspace (owner, step-up, exact name)",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/transfer-ownership": {
      "post": {
        "operationId": "Workspace_transfer",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnershipDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hand the workspace to another member (owner, step-up)",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/me/permissions": {
      "get": {
        "operationId": "Workspace_myPermissions",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyPermissionsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/me/presence": {
      "patch": {
        "operationId": "Workspace_presence",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePresenceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/members": {
      "get": {
        "operationId": "Workspace_listMembers",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberViewDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/members/{userId}": {
      "patch": {
        "operationId": "Workspace_updateMember",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a member (a role change needs a step-up)",
        "tags": [
          "workspaces"
        ]
      },
      "delete": {
        "operationId": "Workspace_removeMember",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/departments": {
      "get": {
        "operationId": "Workspace_listDepartments",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentViewDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "operationId": "Workspace_createDepartment",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/departments/{id}": {
      "patch": {
        "operationId": "Workspace_updateDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "delete": {
        "operationId": "Workspace_removeDepartment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations": {
      "get": {
        "operationId": "Workspace_listInvitations",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvitationViewDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      },
      "post": {
        "operationId": "Workspace_invite",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "8–64 characters of [A-Za-z0-9_-], unique per intended operation",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateInvitationsResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Invite by email or Iranian mobile; each entry is normalised and checked",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations/{id}": {
      "delete": {
        "operationId": "Workspace_revokeInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/roles": {
      "get": {
        "operationId": "Workspace_listRoles",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleViewDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/roles/{roleId}/permissions": {
      "put": {
        "operationId": "Workspace_replacePermissions",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "The role version being replaced",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleViewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace a role’s matrix row (step-up, If-Match, rank and no-escalation rules)",
        "tags": [
          "workspaces"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/roles/reset-defaults": {
      "post": {
        "operationId": "Workspace_resetRoles",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleViewDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "workspaces"
        ]
      }
    }
  },
  "info": {
    "title": "Taskin API",
    "description": "REST API of the Taskin collaboration platform (RFC 0001). Errors are RFC 9457 problem details.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40,
            "example": "هلدینگ راهنما"
          },
          "description": {
            "type": "string",
            "maxLength": 160
          },
          "tone": {
            "type": "string",
            "enum": [
              "brand",
              "teal",
              "violet",
              "amber",
              "rose",
              "slate"
            ]
          },
          "iconUploadKey": {
            "type": "string",
            "description": "The key of a completed icon upload ticket"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "WorkspaceViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "initials": {
            "type": "string"
          },
          "tone": {
            "type": "string",
            "enum": [
              "brand",
              "teal",
              "violet",
              "amber",
              "rose",
              "slate"
            ]
          },
          "iconUrl": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "planId": {
            "type": "string"
          },
          "limits": {
            "$ref": "#/components/schemas/PlanLimitsDto"
          },
          "memberCount": {
            "type": "number"
          },
          "settings": {
            "$ref": "#/components/schemas/WorkspaceSettingsDto"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "initials",
          "tone",
          "iconUrl",
          "ownerId",
          "planId",
          "limits",
          "memberCount",
          "settings",
          "createdAt"
        ]
      },
      "UploadTicketDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "maxBytes": {
            "type": "number"
          },
          "contentTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "expiresInSeconds": {
            "type": "number"
          }
        },
        "required": [
          "key",
          "url",
          "fields",
          "maxBytes",
          "contentTypes",
          "expiresInSeconds"
        ]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "AcceptInvitationResultDto": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "admin",
              "manager",
              "member",
              "guest"
            ]
          }
        },
        "required": [
          "workspaceId",
          "role"
        ]
      },
...
      "DeleteWorkspaceDto": {
        "type": "object",
        "properties": {
          "confirmName": {
            "type": "string",
            "description": "The exact workspace name"
          }
        },
        "required": [
          "confirmName"
        ]
      },
...
    }
  }
}